extracted cea.schemas from cea.scripts
