+++ cea/datamanagement/databases_verification.py
@@ -4,7 +4,7 @@ This tool is used as to check the format of each database
 """
 from __future__ import division
 from __future__ import print_function
-from cea.scripts import schemas
+from schemas import schemas
 import pandas as pd
 import re
